+++ src/blob/core.rs
@@ -348,10 +348,6 @@ where
         self.name.id
     }
 
-    pub(crate) fn offload_filter(&mut self) {
-        self.index.offload_filter()
-    }
-
     pub(crate) async fn check_filters(&self, key: &K) -> Result<bool> {
         trace!("check filters (range and bloom)");
         if let FilterResult::NotContains = self.index.check_filters_key(key).await? {
@@ -524,5 +520,36 @@ impl RawRecords {
         self.current_offset += header.meta_size();
         self.current_offset += header.data_size();
         Ok(header)
+    }
+}
+
+#[async_trait::async_trait]
+impl<K> BloomProvider<K> for Blob<K>
+where
+    for<'a> K: Key<'a> + 'static,
+{
+    type Filter = Bloom;
+    async fn check_filter(&self, item: &K) -> FilterResult {
+        self.index.check_bloom_key(item).await.unwrap_or_default()
+    }
+
+    fn check_filter_fast(&self, item: &K) -> FilterResult {
+        self.index.check_bloom_key_in_memory(item)
+    }
+
+    async fn offload_buffer(&mut self, _: usize, _: usize) -> usize {
+        self.index.offload_filter()
+    }
+
+    async fn get_filter(&self) -> Option<Self::Filter> {
+        Some(self.index.get_bloom_filter().clone())
+    }
+
+    fn get_filter_fast(&self) -> Option<&Self::Filter> {
+        Some(self.index.get_bloom_filter())
+    }
+
+    async fn filter_memory_allocated(&self) -> usize {
+        self.index.get_bloom_filter().memory_allocated()
     }
 }