+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Pearl changelog
 
 ## [Unreleased]
 #### Added
+- Add hierarchical filters support (#126)
 
 
 #### Changed
@@ -13,6 +14,11 @@ Pearl changelog
 
 
 #### Updated
+
+
+## [0.9.0] - 2021-12-09
+#### Added
+- Add hierarchical filters support (#126)
 
 
 ## [0.8.1] - 2021-12-02